Adds an option for the user to add his/her own filters, without having to stick to css hiding rules only... (for example for blocking everything from a domain)
It can be found on the blacklist and whitelist options pages. When you click the button a new tab 'Advanced' is opened.

It just stores the list as a new 'default' list. The onlydifference is that it has no URL (and so isn't modified when you press the update now button).

I tested it as good as I can (and I don't know errors at the moment)...

Strange there isn't a issue for this already. I really thought there already was one...

And now I have programmed enough again, so I am going back to my study books :).