Update issue 5406 Turn GLOBAL_contentScriptData into an object that tracks the number of times it is used, deleting its data after everybody has used it. This lets adblock_start.js and bandaids.js run in arbitrary order while still cleaning up after ourselves.
I was going to make adblock_start use onReady as well, and then have
get_content_script_data just call setData(), but that would cause
adblock_start's code to be run in a setTimeout(0ms) -- aka, it would
put the code at the end of the queue of scripts waiting to run.  Since
we want this code to run as quickly as possible, I didn't go that route.

If you can make GLOBAL_contentScriptData simpler, go for it; I'm just
out of time for today.

Anyway, look at the console.warn messages to watch each piece of code
running and eventually the data being deleted.  We should remove the
.warn()s before shipping.