Update issue 5274 Trying out in a new branch: issue-5274_2 Mind that, when switching to this branch :)
I guess it is impossible then to show the blockable items list in Safari. However, that file has a second function too: create a blocking rule for a single resource. And this shouldn't require any part of the filterset, nor direct calls to the background page.
@John, would you mind to test this out in Safari?
It's about the link to block a single resource. The other one won't be shown. Can you paste the full log output here? I also added some random comments to see where the code breaks if it does.