Update issue 6596 Still needs thorough testing for bugs, but:
Don't register a beforeload handler if not picreplacing
Instead, use querySelectorAll to search for the three or six
possible ways the page could have requested the ad URL.

Initial tests seem to imply that elements are destroyed slightly sooner than in the old approach,
because they flash up on the screen on my test page for less time.

Also, added a tools/tests/ directory with tests for my selector generation code.