Update  Issue 4899 Status: NeedsReview cc: adblockforchrome
Added a function to handle keypress 10 or 13 which is pressing enter button. This handler will then simulate clicking the first button in ui-dialog-buttonset which in turn will invoke the function binded to that button.

To test black list ui, go to msn.com, choose block an ad on page in adblock popup, then instead of clicking the buttons, use enter instead. For white list ui, choose don't run on pages on this domain and use enter instead of buttons.