Update  Issue 4899 Status: NeedsReview cc: adblockforchrome
Encapsulated the binding of default buttons in a function that is called when the call back for load_jquery_ui is invoked. Also, added class for the default buttons for easier binding.

To test choose don't run on pages on this domain and use enter instead of buttons, then instead of clicking the buttons, use enter instead.